Improve class methods handling
